refactor: extract product filters component
draft
